+++ src/Libraries/RevitNodes/Elements/Roof.cs
@@ -146,7 +146,7 @@ namespace Revit.Elements
         /// Create a Revit Roof given its curve outline and Level
         /// </summary>
         /// <param name="outline"></param>
-        /// <param name="RoofType"></param>
+        /// <param name="roofType"></param>
         /// <param name="level"></param>
         /// <returns>The Roof</returns>
         public static Roof ByOutlineTypeAndLevel(Curve[] outline, RoofType roofType, Level level)
@@ -182,9 +182,9 @@ namespace Revit.Elements
         public static Roof ByOutlineExtrusionTypeAndLevel(PolyCurve outline, RoofType roofType, Level level, ReferencePlane plane, double extrusionStart, double extrusionEnd)
         {
 
-            if (!outline.IsClosed)
+            if (outline.IsClosed)
             {
-                throw new ArgumentException(Properties.Resources.OpenInputPolyCurveError);
+                throw new ArgumentException(Properties.Resources.CloseInputPolyCurveError);
             }
 
             var ca = new CurveArray();